style: change actions select style
